+++ frontend/src/pages/EditPage/hooks/useAddTodoBottomSheet.tsx
@@ -1,9 +1,9 @@
 import useAddTodoMutation from './useAddTodoMutation';
-import useBaseBottomSheet from './useBaseBottomSheet';
 import TodoAddForm from '../components/TodoTab/TodoAddForm';
 
 import { TodoCreateParams } from '@/api/todo';
 import { TODO_TOAST_MESSAGE } from '@/constants/message';
+import useBaseBottomSheet from '@/hooks/useBaseBottomSheet';
 import useToast from '@/hooks/useToast';
 
 export const useAddTodoBottomSheet = (dateType: string, planId?: number) => {
