+++ frontend/src/pages/EditPage/hooks/useEditTodoBottomSheet.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 
-import useBaseBottomSheet from './useBaseBottomSheet';
 import useEditTodoMutation from './useEditTodoMutation';
 import TodoEditForm from '../components/TodoEditForm';
 
+import useBaseBottomSheet from '@/hooks/useBaseBottomSheet';
 import { Todo } from '@/types/todo';
 
 const useEditTodoBottomSheet = (dateType: string, planId?: number) => {
